Make the platform a remote execution happened on available to `@rule`s. (#9198)
### Problem

We want to be able to track the actual Platform that an external process was ran on in as a field of a new type `FallibleExecuteProcessResultWithPlatform`. This is important for ensuring that some cross platform workflows (such as, a user running pants on a Mac remoting on a build farm with Linux build machines) work correctly. 

### Solution

This commit creates several new types:
- a Rust `Platform` type analogous to the Python `Platform` type and distinct from the `PlatformConstraint` type that already existed in Rust.
-a new pair of identically-named Python/Rust types meant to be passed over the FFI called `FallibleExecuteProcessResultWithPlatform` which has a new parameter `platform` representing the type of the remote platform, and otherwise contains the fields of the existing `FallibleExecuteProcessResult` type

And rewrites a bunch of `process_execution` code (including test code) to make use of these new types in place of the old `FallibleExecutePlatformResult` type (that is, much of the content of this commit is boilerplate). There are some substantive changes to `cache.rs` to handle correctly storing and reading the Platform of a cached `FallibleExecuteProcessResultWithPlatform`. Finally, this commit renames the string representation of the OSX Platform from "osx" to "darwin", to match what the Python code assumes.

### Result

Rule-authors can request a `FallibleExecuteProcessResultWithPlatform` which has the data representing the remote platform. The existing `FallibleExecuteProcessResult` and `ExecuteProcessResult` types work the same way as they previously did, but are now ultimately derived from `FallibleExecuteProcessResultWithPlatform` by dropping the platform information.