Copy jar.create products to the dist directory
The `jar.create` task has limited usability because its product is
created deep under the .pants.d directory. Certain workflows may benefit
from having easier access to the jar file (e.g. for publishing it as a
binary artifact).

This commit improves JarCreate so that it copies its products to the
`dist` directory each time the task is executed.