Targets store all their fields with their unhydrated values
We must store all the unhydrated values at parse time. Once we parse a BUILD file, we never want to re-parse it, so we instead store all that information in the most minimally hydrated form possible.

We also calculate all default values when the fields are not explicitly defined.

We add Target.get(Type[Field]) -> Field as a way to get MyPy working with the target adaptor. Through generics, MyPy will know exactly which Field was returned!

The definition of Target is now mostly complete!