Redesign `test.py` to work with custom target types (#9276)
### Problem

One of the goals of the upcoming Target API is to provide extensibility so that V2 works with custom target types. For example, we should still run tests with a custom `python2_tests`/`python3_tests` target. 

Right now, `test.py` is organized around the idea of `TestTarget`, which is a union with members like `PythonTestsAdaptor` or `JvmTestsAdaptor`. This means that we expect each test runner implementation to only work with one single target type. So, it would be impossible with the current design to have `python2_tests` work with `python_test_runner.py`. 

Instead, we need a way for a test runner like `pytest_runner.py` to work with multiple target types.

### Solution

Instead of `UnionRule(TestTarget, PythonTestsAdaptor)`, use `UnionRule(TestRunner, PytestRunner)`. 

Each `TestRunner` union member has a method `TestRunner.is_valid_target()`, which determines if the test runner can meaningfully operate on that target or not. Currently, `PytestRunner.is_valid_target()` checks that the target is an instance of `PythonTestsAdaptor`. Once we have the Target API, we will instead rewrite this to check if the target has the required typed `Field`s. It's crucial that this check happens through standard Python rather than union membership, as it makes the implementation of the Target API much simpler: https://docs.google.com/document/d/1nxPdvuzgCPKhTabhfYBN2tbcr85enSR83OpcvdfWXfY/edit#heading=h.z97og7gj9rvs.

This mirrors the changes we made to `lint` and `fmt` in https://github.com/pantsbuild/pants/pull/9207 and https://github.com/pantsbuild/pants/pull/9211.

### Result

`python_test_runner.py` can now work with multiple target types once we have the Target API.

A change in behavior is that we now always no-op when the specified target is not a valid test target, whereas we previously would error if a single address or single file arg was used; this is the same behavior as `fmt` and `lint`. We can add back that error behavior if we want to.