Allow subclassing `Field`s for custom `Field` behavior with the Target API (#9286)
### Problem

One of the requirements for the Target API's extensibility goal is that plugin authors can create new target types that behave almost identically to the pre-existing target type, but that somehow override a field. For example, `python3_library` may override the `Compatibility` field to always return `CPython>=3.5`.

(We don't allow plugins to change the behavior of a pre-existing target. You must create a new target type.)

Plugin authors may subclass a particular `Field` currently, but then core Pants code will stop working correctly. For example, Core Pants code may filter out targets by the `Compatibility` field. So, a `Python3Compatibility` field that subclasses `Compatibility` would no longer be recognized when calling `my_tgt.has_fields([Compatibility])` or `my_tgt.get(Compatibility)`, as `Python3Compatibility != Compatibility`, even though it is a subclass.

### Solution

Modify `Target.get()` and `Target.has_fields()` to first check if the exact type is registered on the Target, then fall back to checking if the requested Field is a superclass of any of the registered fields.

### Result

A plugin author can now define `Python3Library`, which is identical to `PythonLibrary`, except for having `Python3Compatibility` in lieu of `Compatibility`. Code that depends on `my_tgt.has_fields([Compatibility])` will continue to work with this custom target type.

If a plugin needs the more precise `Field`, that will work too. `my_tgt.has_fields([Python3Compatibility])` would return `True` for `Python3Library`, but not `PythonLibrary`. This is a useful property—we can, for example, filter to get all targets with `PythonSources`, even if other targets have the field `Sources`.