Change `UnionMembership` to use `dict` instead of `OrderedDict` (#9288)
As of Python 3.6+, dictionaries are ordered by default. (While this is not formalized until Python 3.7, the implementation detail is guaranteed in 3.6 and it was not formalized in 3.6 so that later Python versions maintained flexibility to change the guarantee.)

Dicts are more ergonomic than `OrderedDict`s, such as being easier to create and having a simpler `repr()`.