Tweak how to add plugin fields to V2 targets to be more explicit (#9291)
Unions are supposed to be an `is-a` relationship.

Previously, to add a new field to a pre-existing target, you would do `UnionRule(PythonLibrary, TypeChecked)`. This is not an `is-a` relationship, though; `TypeChecked(Field)` is not a `PythonLibrary(Target)`.

Instead, we add a sentinel class automatically to every `Target` called `PluginField`. So, to add a new field, plugin authors simply register `UnionRule(PythonLibrary.PluginField, TypeChecked)`.