Make Target API safe to use with the engine (#9303)
### Problem

The engine requires that every value is immutable and hashable. (Technically, it only enforces hashability, but mutable things are fundamentally not safe to hash).

However, many of the `raw_value`s we stored on `Field` are not hashable, such as `List`. The idea of `raw_value` was to preserve exactly what the user typed in a `BUILD` file, but that's fundamentally not safe to do because `BUILD` files can use mutable data types.

### Solution

For `PrimitiveField`, discard the `raw_value` after initialization. We were only keeping it for the sake of debugging. It would be neat to continue storing, but it's not safe to do so with the engine so the best option is to simply discard the value. This also improves the space usage of `PrimitiveField`.

For `AsyncField`, due to lazy hydration, we must preserve the raw value. But, we need to use these fields with the engine, so we introduce the attribute `sanitized_raw_value` and the abstractmethod `sanitize_raw_value()`. This allows us to do things like convert lists into tuples. `sanitize_raw_values()` also means that we can now get some light-weight validation to be eager, which is likely a good thing as it allows us to catch invalid BUILD files eagerly. For example, we can eagerly fail if a user tries doing `sources=[0, 1, 2]`.

### Result

`Target` and `Field` are now safe to use with the engine.*

*`Field` authors can still mess things up. They must use immutable and hashable types for `PrimitiveField.value` and `AsyncField.sanitized_raw_value`. There is nothing enforcing a `Field` author to use a safe type like a tuple over a list, other than convention.** 

** Maybe we could make a MyPy plugin to check that the return type of a function is immutable? This would need to work ergonomically, such as working with native tuples and not forcing us to subclass `Tuple`.