Switch to a dict-backed OrderedSet impl. (#9305)
### Problem

#9181 caused a significant performance regression for operations in larger repos. It was also noticeable in the pantsbuild/pants repo:
```
# for 5 runs of `./pants filter :: | wc -l`:

# before #9181 
0m33.455s
# after #9181 
1m23.784s
```

Profiling showed that all of the time increase was in the `__contains__` method, which in #9166 had gone from `O(1)` to `O(n)`. The benchmark done there to justify this likely used 1) too small a set, 2) with too simple an `__eq__` impl ... to identify the difference.

### Solution

Switch the backing datastore to a python `dict`. Since pants requires python>=3.6, `dict` guarantees order. `*OrderedSet` `__eq__` and `__hash__` receive custom but straightforward implementations. Additionally, mixing-in of `Sequence[T]` and indexing/slicing methods were removed, since they cannot be implemented efficiently with `dict` as it stands, and were as-of-yet unused.

### Result

Performance is restored to before the change; code is slightly simpler.