A newtype for the requirements.txt file dep.
We don't want to invalidate sources based on arbitrary
changes to requirements.txt, but we do need to depend on
it so we know to recompute requirements.

Previously this was achieved by synthesizing a files()
target and adding a dep on it, in the python_requirements macro.
However when we gather sources to operate on we have to include
all files(), meaning we were invalidating the sources even on
a whitespace change to requirements.txt.

This change adds a new, internal, target type specially for
the synthetic requirements.txt change, so that we can maintain
a dependency, but rules can still filter out the actual file.

Context: https://github.com/pantsbuild/pants/pull/6405/