Add `Target.has_field()` as sugar for `Target.has_fields()` (#9314)
The main idiom we'll use for Target API code is:

```python
relevant_tgts = [tgt for tgt in tgts if tgt.has_field(PythonSources)]
# do something with these tgts
```

`has_field(PythonSources)` is simpler to write than `has_fields([PythonSources])`. Because we will use this so often, this is nice sugar to add.

--

This also documents the 3 main methods exposed by `Target`.