Allow registering new target types with the Target API (#9334)
## Problem 1: registering new target types

We need a light-weight mechanism for plugin authors to register new target types in their `register.py`.

In V1, you do this:

```python
def build_file_aliases():
    return BuildFileAliases(
        targets={
            PythonBinary.alias(): PythonBinary,
            PythonLibrary.alias(): PythonLibrary,
            PythonTests.alias(): PythonTests,
        },
```

Instead, the Target API registration is much lighter weight:

```python
def targets2():
    return [PythonBinary, PythonLibrary, PythonTests]
```

We use `targets2()` to retain flexibility in the case that we want to change this endpoint.

It's not yet clear how `objects` and `context_aware_object_factories` should be registered. That is punted on for now. Those are much less common than registering a `Target`, so the registration of a `Target` should not be made more difficult/clunky to accommodate `objects`.

### Deferred problem: registering the alias with the Parser

In a followup, we will replace the explicit registration of symbols like `python_library` in `engine_initializer.py` with instead dynamic registration of symbols based on all the registered Target types.

## Problem 2: Allow rules to request all registered target types

We need a singleton for rules to get all the registered target types. For example, we need this to  convert a `HydratedStruct -> Target` so that we know which `Target` subclass to use.

A simple singleton would return `FrozenOrderedSet[Type[Target]]`. But, a very common pattern is to want to be able to quickly map from `alias -> target type`; for example, we need this mapping to efficiently know how to convert `HydratedStruct` into a `Target` subclass.

So, we add a new `RegisteredFieldTypes` dataclass which is a light wrapper around a dictionary of `alias -> target type`. We use properties to allow call sites to very easily (and performantly) get all aliases and/or get all target types.

## Result

We have a new rule to go from `HydratedStruct -> Target`. `python_create_binary.py` now exclusively uses the engine to go from `Address -> PythonBinary -> PythonBinaryFields`.