Make LocalPantsRunner a dataclass (#9339)
This commit attempts to make the code in LocalPantsRunner cleaner by turning it into a dataclass, which allows us to get rid of the `__init__` method. Some single-use methods and properties on `LocalPantsRunner` have also been made more concise, and a number of type annotations were added to it and in several other places. `local_pants_runner.py` is now about 30 lines shorter