Report warnings and errors to reporting server for JVM targets (#9293)
At Twitter, we are interested in monitoring the number of warnings triggered by each target to gain better insight into our development practices.
We want for the solution to be effective regardless of whether we compiled  the code at  this very moment or we fetched the compilation output from a local or remote cache.

* Update zinc to the latest version which supports reporting diagnostics data
* Have zinc report the data about diagnostics to a json file in the buildcache
* Send aggregated data about diagnostics for jvm targets to our reporting servers using the `target_scope_data` API
* To report the data also when the cache was hit, add a hook function to the `JvmCompile` base class
* Feature gate the reporting behind the advanced user facing opiton: `--compile-rsc-report-diagnostic-counts` which is off by default

For each target, the following additional data is communicated to our reporting server:
```
n_diagnostics.Errors,
n_diagnostics.Warnings,
n_diagnostics.Information,
n_diagnostics.Hint,
```
That is, the count of diagnostics in this target for each severity as defined by the Language Server Protocol.