Skip Rust tests and JVM tests in CI when no respective changes were made (#9367)
## Problem

Our CI is painfully slow and flaky. 

One easy win we can make is to stop running irrelevant jobs when there is no need to. We already do this for docs-only changes, where we skip all of CI.

One other low hanging fruit is to skip any Rust tests and JVM tests when no changes were made to their relevant files.

## Solution

Use Travis's built-in skip mechanism https://docs.travis-ci.com/user/conditional-builds-stages-jobs#conditional-jobs to skip Rust and JVM when possible. Per its [Conditions docs](https://docs.travis-ci.com/user/conditions-v1), we can't use complex logic to calculate the skip, e.g. we can't run some script. So, we instead inspect the commit message to see if it has the special lines `[ci skip-rust-tests]` and/or `[ci skip-jvm-tests]`. 

Specifically, we teach the githook `prepare-commit-msg` to not only look for docs-only changes, but to also look if Rust-related files are changed and JVM-related files are changed.

### Checking the entire changeset rather than the last commit

Our previous implementation was buggy in that it would only check for the most recent commit. For example, even if you had a commit with 20 Python files changed, if the next commit only updated a doc, then we would skip CI for the whole PR.

Instead, we check against the merge base for the files changed, which typically corresponds to the changes since the most recent push to the remote / PR.

### Also extract `calculate_engine_hash.sh`

This refactor makes it easier to keep the githook in sync with the `native_engine.so` fingerprint code. This also is prework for a followup change to start caching `native_engine.so` in CI.

## Result

The majority of builds will have 4 fewer shards and save about 40 minutes of overall CI time.

If users want to override this automatic skipping, they can easily modify the commit message to remove the `[ci skip-{rust,jvm}-tests]` label.