Improve console output of binary and awslambda goals.
Previously the output was:

```
Generating binaries in `./dist`
Wrote /full/path/to/dist/helloworld.pex
```

The two lines were inconsistent (one used the relpath and
the other a fullpath), and the first is redundant anyway.

Now this outputs:

```
Wrote dist/helloworld.pex
```

Which retains all the information, but is more succinct.

[ci skip-rust-tests]  # No Rust changes made.

[ci skip-jvm-tests]  # No JVM changes made.