Finish the tests using stricter strict equality checks
This allows us to ensure we really are testing this the right way. The tests are so lightweight and are 100% self-contained, so this is worth the increased fragility. This allows us to enforce that the tests are really doing everything we care about:

* sorting
* adding leading and trailing whitespace correctly
* handling new lines correctly

This is like having an integration test.

# Delete this line to force CI to run Clippy and the Rust tests.
[ci skip-rust-tests]  # No Rust changes made.

# Delete this line to force CI to run the JVM tests.
[ci skip-jvm-tests]  # No JVM changes made.