Add `FrozenDict` as an immutable wrapper around native dictionaries (#9366)
## Problem

We have no ergonomic way to work with dictionaries with the engine, as `dict` is mutable and not hashable. We've resorted to using custom constructors in dataclasses that allow the user to pass a `Dict`, and then, in the constructor, converting it to a `Tuple[Tuple[K, V], ...]`, which is still clunky to use.

https://github.com/pantsbuild/pants/blob/eeb9d7488a39bb428bf19e273584675d1041745b/src/python/pants/engine/isolated_process.py#L24-L62

## Solution

Add `FrozenDict` as a very lightweight wrapper around native `dict`, with the only differences that it removes all entry points for mutation (e.g. `d.update()`) and it implements `__hash__()`.

### Rejected alternative: proactive casting of unhashable elements

Toolchain has a `FrozenDict` implementation that proactively attempts to convert hashable elements to hashable, and has additional guarantees to ensure the dict truly is frozen. For example, it will convert `{'a': [1]}` to `{'a': (1,)}`.

Instead, we take a much simpler approach to validate and give a nice error message when using unhashable elements, so that the rule author may fix their code. This is simpler, more performant, and more transparent.

### Does not use `typing.Hashable`

This is a way in MyPy to express that something must must be hashable. However, it does not work as well as expected, e.g. complaining about `typing.Type`, such as `str`, not being hashable, even though `hash(str)` works.

## Result

In a followup, we will audit the current engine code to see where we can clean things up by using `FrozenDict`.