Formalize default values for Fields (#9384)
Currently, we allow subclasses of `BoolField` and `Sources` to specify a default value, meaning a value used when the user did not explicitly specify the field in a BUILD file (i.e. `raw_value=None`). Everything else is hardcoded to default to `None`.

By formalizing that every field, not just those two, has a default field, we get these benefits:

1. Easier extensibility for subclassed fields. For example, a `StringField` subclass may want to keep the string type checking but change the default from `None` to `my_str`. Now that is possible.
2. Better output in `./v2 target-types2`. We can now always give the default value:

![Screen Shot 2020-03-24 at 12 17 17 AM](https://user-images.githubusercontent.com/14852634/77398469-d3746800-6d64-11ea-8519-7a4de2c801ac.png)

Note that the vast majority of fields will still default to `None`, which we want. But, subclasses can modify this as desired.


[ci skip-rust-tests]  # No Rust changes made.
[ci skip-jvm-tests]  # No JVM changes made.