Revert internally enforcing Flake8's E231 trailing comma lint (#9387)
For example, we will once more allow `['foo',]`, rather than requiring `['foo']`.

While these trailing commas are hard to read, Black does not automatically remove them. When deciding to use Black, we ceded control of formatting to the tool, even though that means ~5% of its decisions we won't like. It's a bad user experience to say "Black will handle it all, expect for this one thing that you need to manually fix."