Add `SKIP_NATIVE_ENGINE_SO_BOOTSTRAP` override
I originally tried to set up the environment in such a way that we tricked bootstrap_code.sh into not needing to recompile. This ended up being really tricky to do. Not only do we need to have native_engine.so saved in a global file, but we need a metadata file associated with it.

# Delete this line to force CI to run the JVM tests.
[ci skip-jvm-tests]  # No JVM changes made.