Upgrade to Pex 2.1.7.
This brings in fixes that support our use of `--disable-cache` with the
Pex PEX binary that had regressed in the Pex 2.x series.

Also leverage the new `--no-strip-pex-env` option when building the
pants PEX we deploy.

Further improvements including better error messages in some cases are
detailed in the release notes:
  https://github.com/pantsbuild/pex/releases/tag/v2.1.7

Fixes #9361

[ci skip-jvm-tests]  # No JVM changes made.