Fix use of pid (#9400)
In #9389, we removed the `pid` field from `NailgunError`. However, some error-handling code in `remote_pants_runner.py` was still assuming this field existed, and using it to print an additional error message in one case. Now that we don't care about a remote PID in Nailgun code, this error print line should be removed.