Move Build Wheels shards lower in CI order (#9383)
Generally, we order CI based on the likeness to fail - we put things likely to fail closer to the top, e.g. `lint`, so that we get faster feedback.

The Build Wheels shards very rarely fail, but we kept them closer to the top for the sake of releases and for Twitter needing the shards to run to build internal wheels. While both reasons are important, when someone needs the build wheels shards to run first, they can simply cancel all the other shards in the build to remove the queue.

Because there is a viable workaround, the special case of releases + Twitter builds is not enough to justify putting two very stable shards above the V1 test runner integration tests, which do often fail due to changes made in PRs.

# Delete this line to force CI to run Clippy and the Rust tests.
[ci skip-rust-tests]  # No Rust changes made.

# Delete this line to force CI to run the JVM tests.
[ci skip-jvm-tests]  # No JVM changes made.