Make the list of JS lintable extensions configurable (#9423)
### Problem

Node plugin only lints .js/.jsx files. These extensions are hardcoded. However, there are more potentially eslintable extensions, such as .ts or .vue.

### Solution

The list of extensions (or rather suffixes, to not mandate the dot) is now configurable as `--lint-javascriptstyle-file-extensions` (`lint_javascriptstyle` scope). The default value has the same two extensions as before, so the default behaviour is unchanged.

[ci skip-rust-tests]  # No Rust changes made.
[ci skip-jvm-tests]  # No JVM changes made.