Better namespace field class names
Borja pointed out in the Go bindings PR that this will make call sites clearer which field they are referring to. Go already has a `Packages` field (now called `GoPackages`), so `Packages` is confusing to also use in `backend/native` and `ConanPackages` is more descriptive. The BUILD file alias remains the same.

# Delete this line to force CI to run Clippy and the Rust tests.
[ci skip-rust-tests]  # No Rust changes made.

# Delete this line to force CI to run the JVM tests.
[ci skip-jvm-tests]  # No JVM changes made.