Add `convert_source_to_sources.py` script to help users migrate away from deprecated `source` field (#9309)
We will soon be deprecating `source` in favor of `sources`. 

This is a highly tedious and automatable thing to fix, so we will provide a script to fix ~95-100% of usages. (This handles almost every edge case, except for weird patterns like using a `function` in the `source` argument; it even handles variables.)

We also run the script on the internal codebase to prepare for the change.

[ci skip-rust-tests]
[ci skip-jvm-tests]