Fix targets that do not have a `dependencies` field with the Target API  (#9481)
### Problem

Several targets do not have the `dependencies` field because it does not make sense for them. For example, in V1, it is an error to define `dependencies` on an `alias` target:

> ERROR: Build graph construction failed: ExecutionError 1 Exception encountered:
  MappingError: Failed to parse src/python/pants/util/BUILD:
TypeError("create_object() got multiple values for keyword argument 'dependencies'",)

Likewise, it does not make sense to allow `dependencies` for the private target type `_python_requirements_file`.

--

However, `TargetAdaptor` subclasses `StructWithDeps`, which we use to resolve `dependencies` to actual `Addresses`. This means that every single `HydratedStruct`/`TargetAdaptor` will have a `dependencies` value in its `struct.kwargs()`, even if the user left off the field.

So, we then try passing `dependencies=None` to the `Target` constructor and it fails that it's an unrecognized field for that target type.

### Solution

In the `HydratedStruct -> Target` code, explicitly check if the field does not register `Dependencies` (or a subclass) and discard the `dependencies=None` if so.

[ci skip-rust-tests]
[ci skip-jvm-tests]