[IDE] Precompute flattened dependencies for modulizable targets (#9311)
Currently, we walk the transitive closure of every (non-strict-deps) target twice: Once to figure out what dependencies it should have in the classpath, and another one to figure out which of those should be made into libraries. We do that for every modulizable target, which makes export-dep-as-jar perform non-linearly with respect to the number of modulizable targets.

NOTE: I think this is a pretty convoluted graph thing, so I would love some input on whether this is readable and understandable, and how to improve that if not.

We essentially need to be able to do three things with the transitive dep graph of a target:

Calculate the libraries entry: This the flat sorted list of transitive dependencies that are not modulizable targets, respecting strict deps.
Calculate which modulizable targets are direct dependencies of mine, for the targets field.
Calculate flat closure of my transitive dependencies, modulizable and not modulizable, for source_dependencies_in_classpath, respecting strict deps.
So, we implement it this way:

Traverse the whole dependency graph once to create a Map[Target, Set[NonModulizableTarget]], which will contain the flat list of non-modulizable dependencies. The libraries field for a target will correspond exactly with an entry in this map.
Pass that map to the processing of every target. The entry in source_dependencies_in_classpath will be the union of all the entries of all the transitive modulizable dependencies of a target. This operation should be cheap, or at least scale linearly with the size of modulizable_target_set.
E.g. if A -(depends on)-> B -> C -> D and A -> E and B -> F, if modulizable_targets = {A, B}, the resulting map will be:

   flat_deps =  {
           A -> {E},
           B -> {C, F, D},
           // Some other entries for intermediate dependencies
           C -> {D},
           E -> {},
           F -> {},
        }
From here:

targets of A: Computed trivially from A.dependencies.
libraries of A: Computed from the map (libraries = flat_deps[A])
source_dependencies_in_classpath of A: Computed by traversing the dep closure and unioning the entries for modulizable deps ( source_deps_in_classpath = flat_deps[A] U flat_deps[B], but not flat_deps[C], because C is not modulizable).

No observable functionality impact, but a performance improvement.